requirements.txt
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
requests==2.32.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.33.2